Moved User to Server
Moved the User class to the Server folder, updated references as well
